Issue-1121: Rename the org.carlspring.strongbox.storage.metadata.nuget.metadata as to not repeat itself
