+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-nuget-layout-provider/src/main/java/org/carlspring/strongbox/storage/metadata/nuget/Framework.java
@@ -15,14 +15,13 @@
  * limitations under the License.
  */
 
-package org.carlspring.strongbox.storage.metadata.nuget.metadata;
-
-import static java.text.MessageFormat.format;
+package org.carlspring.strongbox.storage.metadata.nuget;
 
 import java.util.EnumSet;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import static java.text.MessageFormat.format;
 
 /**
  * Represents .net framework of a nuspec field.
