+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-nuget-layout-provider/src/test/java/org/carlspring/strongbox/storage/metadata/nuget/metadata/FrameworkAssemblyTest.java
@@ -17,12 +17,12 @@
 
 package org.carlspring.strongbox.storage.metadata.nuget.metadata;
 
+import org.carlspring.strongbox.storage.metadata.nuget.AssemblyTargetFrameworkAdapter;
+import org.carlspring.strongbox.storage.metadata.nuget.Framework;
+
 import java.util.EnumSet;
 
-import org.carlspring.strongbox.storage.metadata.nuget.metadata.AssemblyTargetFrameworkAdapter;
-import org.carlspring.strongbox.storage.metadata.nuget.metadata.Framework;
 import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.assertNull;
 
 /**
