+++ strongbox-storage/strongbox-storage-layout-providers/strongbox-storage-nuget-layout-provider/src/test/java/org/carlspring/strongbox/storage/metadata/nuget/metadata/VersionRangeTest.java
@@ -17,13 +17,11 @@
 
 package org.carlspring.strongbox.storage.metadata.nuget.metadata;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import org.carlspring.strongbox.storage.metadata.nuget.VersionRange;
 
-import org.carlspring.strongbox.storage.metadata.nuget.metadata.VersionRange;
 import org.junit.jupiter.api.Test;
 import org.semver.Version;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  *
